requirements.txt
fiftyone==1.5.2
ipython==8.12.3
lsnms==0.4.4
numpy==1.24.4
onnxruntime==1.15.0
onnxruntime_gpu==1.15.0
opencv_python==4.8.0.76
opencv_python_headless==4.10.0.84
Pillow==11.2.1
pycocotools==2.0.10
Requests==2.32.4
scikit_image==0.19.3
timm==0.9.7
torchvision==0.15.2+cu117
ultralytics==8.3.97
